Fix bug with progress bar (reported by Serj_K)
git-svn-id: http://svn2.assembla.com/svn/secu3man@220 c1dae7de-a428-49f4-85d8-50b65f41d01d

+++ sources/secu3man/MainFrame/StatusBarManager.h
@@ -39,7 +39,7 @@ class CStatusBarManager
   void SetInformationText(const CString& i_text);
 	
   void ShowProgressBar(bool show);
-  void SetProgressRange(short nLower, short nUpper);
+  void SetProgressRange(int nLower, int nUpper);
   void SetProgressPos(int nPos);
   void SetLoggerState(int i_state);
